refactor: move injection utils to remark pkg
