+++ packages/astro/src/@types/astro.ts
@@ -1464,10 +1464,6 @@ export interface SSRResult {
 	_metadata: SSRMetadata;
 }
 
-export type MarkdownAstroData = {
-	frontmatter: MD['frontmatter'];
-};
-
 /* Preview server stuff */
 export interface PreviewServer {
 	host?: string;
