+++ packages/astro/src/vite-plugin-utils/index.ts
@@ -1,6 +1,5 @@
 import ancestor from 'common-ancestor-path';
-import type { Data } from 'vfile';
-import type { AstroConfig, MarkdownAstroData } from '../@types/astro';
+import type { AstroConfig } from '../@types/astro';
 import {
 	appendExtension,
 	appendForwardSlash,
@@ -34,32 +33,6 @@ export function getFileInfo(id: string, config: AstroConfig) {
 		fileUrl = appendExtension(fileUrl, 'html');
 	}
 	return { fileId, fileUrl };
-}
-
-function isValidAstroData(obj: unknown): obj is MarkdownAstroData {
-	if (typeof obj === 'object' && obj !== null && obj.hasOwnProperty('frontmatter')) {
-		const { frontmatter } = obj as any;
-		try {
-			// ensure frontmatter is JSON-serializable
-			JSON.stringify(frontmatter);
-		} catch {
-			return false;
-		}
-		return typeof frontmatter === 'object' && frontmatter !== null;
-	}
-	return false;
-}
-
-export class InvalidAstroDataError extends TypeError {}
-
-export function safelyGetAstroData(vfileData: Data): MarkdownAstroData | InvalidAstroDataError {
-	const { astro } = vfileData;
-
-	if (!astro || !isValidAstroData(astro)) {
-		return new InvalidAstroDataError();
-	}
-
-	return astro;
 }
 
 /**
