+++ packages/markdown/remark/src/index.ts
@@ -8,7 +8,7 @@ import rehypeIslands from './rehype-islands.js';
 import rehypeJsx from './rehype-jsx.js';
 import toRemarkContentRelImageError from './remark-content-rel-image-error.js';
 import remarkEscape from './remark-escape.js';
-import { toRemarkInitializeAstroData } from './remark-initialize-astro-data.js';
+import { toRemarkInitializeAstroData } from './frontmatter-injection.js';
 import remarkMarkAndUnravel from './remark-mark-and-unravel.js';
 import remarkMdxish from './remark-mdxish.js';
 import remarkPrism from './remark-prism.js';
