+++ packages/markdown/remark/src/internal.ts
@@ -1,0 +1,5 @@
+export {
+	InvalidAstroDataError,
+	safelyGetAstroData,
+	toRemarkInitializeAstroData,
+} from './frontmatter-injection.js';
